+++ Runtime/Scripts/DecodeResult.cs
@@ -8,29 +8,42 @@ namespace Draco
     /// <summary>
     /// Holds the result of the Draco decoding process.
     /// </summary>
-    public struct DecodeResult
+    public readonly struct DecodeResult
     {
         /// <summary>
         /// True if the decoding was successful
         /// </summary>
-        public bool success;
+        public readonly bool success;
 
         /// <summary>
         /// Axis aligned bounding box of the mesh/point cloud.
         /// </summary>
-        public Bounds bounds;
+        public readonly Bounds bounds;
 
         /// <summary>
         /// True, if the normals were marked required, but not present in Draco mesh.
         /// They have to get calculated.
         /// </summary>
-        public bool calculateNormals;
+        public readonly bool calculateNormals;
 
         /// <summary>
         /// If the Draco file contained bone indices and bone weights,
         /// this property is used to carry them over (since MeshData currently
         /// provides no way to apply those values)
         /// </summary>
-        public BoneWeightData boneWeightData;
+        public readonly BoneWeightData boneWeightData;
+
+        public DecodeResult(
+            bool success,
+            Bounds bounds,
+            bool calculateNormals,
+            BoneWeightData boneWeightData
+            )
+        {
+            this.success = success;
+            this.bounds = bounds;
+            this.calculateNormals = calculateNormals;
+            this.boneWeightData = boneWeightData;
+        }
     }
 }